added auto to cpp_static_analysis so we can remove cpp_goto_auto
