--- sample_files/sample_assignment_config/cpp_goto_auto/config.json
@@ -1,41 +1,0 @@
-{
-    // FIXME:  This section of resource limits should go away soon
-    "resource_limits" : {
-        "RLIMIT_CPU" : 60,
-        "RLIMIT_NPROC" : 100,
-        "RLIMIT_AS" : "RLIM_INFINITY"
-    },
-
-    "testcases" : [
-        {
-            "title" : "Check for goto and auto",
-            "command" : [ "submitty_count_token -l c goto *.cpp",
-                          "submitty_count_token -l c auto *.cpp" ],
-           "points" : 2,
-            "validation" : [
-                {
-                    // number of goto must be 0
-                    "method" : "intComparison",
-                    "actual_file" : "STDOUT_0.txt",
-                    "description" : "Number of `goto`",
-		    "comparison" : "eq",
-                    "term" : 0,
-                    "failure_message" : "You must not use the `goto` keyword.",
-                    "show_message" : "on_failure",
-                    "show_actual" : "never"
-                },
-                {
-                    // number of auto must be 0
-                    "method" : "intComparison",
-                    "actual_file" : "STDOUT_1.txt",
-                    "description" : "Number of `auto`",
-		    "comparison" : "eq",
-                    "term" : 0,
-                    "failure_message" : "You must not use the `auto` keyword.",
-                    "show_message" : "on_failure",
-                    "show_actual" : "never"
-                }
-            ]
-        }
-    ]
-}
