--- sample_files/sample_submissions/cpp_goto_auto/buggy.cpp
@@ -1,5 +1,0 @@
-int main() {
-foo:
-	auto x = 1;
-	goto foo;
-}
