--- sample_files/sample_submissions/cpp_goto_auto/correct.cpp
@@ -1,2 +1,0 @@
-int main() {
-}
