+++ tests/integrationTests/tests/tutorial_05_cpp_static_analysis/validation/results_grade.txt_buggy3
@@ -1,0 +1,4 @@
+Testcase  1: Check for goto and auto                              1 /  2  
+Testcase  2: Submission Limit                                             
+Automatic grading total:                                          1 /  2
+Non-hidden automatic grading total:                               1 /  2
