+++ tests/integrationTests/tests/tutorial_05_cpp_static_analysis/validation/results_grade.txt_solution
